pyproject.toml
[tool.poetry]
name = "advantage20"
version = "0.1.0"
description = ""
authors = ["BindonWrathbite <165421910+BindonWrathbite@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.4"
djangorestframework = "^3.15.2"
django-rest-knox = "^5.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
